requirements.txt
fastapi==0.115.6
python-multipart>=0.0.9
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
greenlet>=3.0.0
asyncpg==0.30.0
pgvector==0.3.6
alembic==1.14.1
sentence-transformers==3.3.1
httpx==0.28.1
python-telegram-bot==21.9
pydantic-settings==2.6.1
google-generativeai==0.8.3
structlog==24.4.0
pytest==8.3.4
pytest-asyncio==0.24.0
ruff>=0.8.0
mypy>=1.13.0
types-python-dateutil
